+++ src/js/core/cover.js
@@ -1,16 +1,6 @@
-import { attr, css, Dimensions, isTag, parent } from 'uikit-util';
+import { css, Dimensions, isTag, parent } from 'uikit-util';
 import { resize } from '../api/observables';
 import Video from './video';
-
-const coverProps = {
-    top: 0, // resets the css from [uk-cover]
-    left: 0, // resets the css from [uk-cover]
-    width: '100%',
-    height: '100%',
-    transform: 'none',
-    objectFit: 'cover',
-    objectPosition: 'center',
-};
 
 export default {
     mixins: [Video],
@@ -37,16 +27,6 @@ export default {
 
     connected() {
         this._useObjectFit = isTag(this.$el, 'img', 'video');
-
-        if (this._useObjectFit) {
-            css(this.$el, coverProps);
-        }
-    },
-
-    disconnected() {
-        if (this._useObjectFit) {
-            attr(this.$el, 'style', '');
-        }
     },
 
     update: {